requirements.txt
Django==5.2.4
Faker==37.5.2
python-dateutil==2.9.0